commit/717
스킬데이터 추가변경
몬스터데이터 추가변경
웨폰데이터 추가변경
베틀매니저 생성
(나중에 매니저분리해야함)
배틀용병 생성
배틀몬스터 생성

+++ Assets/01.Scripts/Mercenary/MercenaryData.cs
@@ -14,8 +14,8 @@ public class Mercenary
     public int exp;
     public int statpoint;
 
-    public int hp;
-    public int maxHp;
+    public float hp;
+    public float maxHp;
 
     public int str;
     public int intel;
